requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Cerberus==1.3.2
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
